Refactor heap allocator to support mmap fallback for large allocations
- Removed custom allocator implementation in `libzip` and replaced it with a macro for `libheap::dll_allocator`.
- Enhanced the heap allocator in `stdlib` to use `mmap()` for large allocations (≥ 64 KiB) to prevent fragmentation of the sbrk heap.
- Introduced a new mechanism to allocate memory via `mmap()` when the sbrk region is exhausted, ensuring efficient memory management.
- Updated allocation and deallocation logic to handle both small (sbrk) and large (mmap) allocations appropriately.
- Added checks to ensure that mmap'd memory is within the expected virtual address range.

+++ libs/libfont/src/syscall.rs
@@ -87,9 +87,9 @@ fn syscall3(num: u32, a1: u64, a2: u64, a3: u64) -> u64 {
 }
 
 /// Extend the process heap by `increment` bytes. Returns previous break address.
-/// Returns `usize::MAX` on failure.
-pub fn sbrk(increment: i32) -> usize {
-    syscall1(SYS_SBRK, increment as u64) as usize
+/// Returns `u64::MAX` on failure.
+pub fn sbrk(increment: u32) -> u64 {
+    syscall1(SYS_SBRK, increment as u64)
 }
 
 /// Open a file. Returns file descriptor or u32::MAX on error.
