+++ libs/libimage/src/syscall.rs
@@ -45,8 +45,9 @@ fn syscall3(num: u32, a1: u64, a2: u64, a3: u64) -> u64 {
 }
 
 /// Extend the process heap by `increment` bytes. Returns previous break address.
-pub fn sbrk(increment: i32) -> usize {
-    syscall1(SYS_SBRK, increment as u64) as usize
+/// Returns `u64::MAX` on failure.
+pub fn sbrk(increment: u32) -> u64 {
+    syscall1(SYS_SBRK, increment as u64)
 }
 
 /// Write bytes to a file descriptor.
